Mark extra sunpy.io readers as private API
