+++ docs/whatsnew/5.0.rst
@@ -36,3 +36,9 @@ The minimum required versions of these optional dependencies has also been updat
 The ``sunpy.database`` module is no longer actively maintained and has a number of outstanding issues.
 It has now been formally deprecated, and will be removed in sunpy 6.0.
 If you are using sunpy.database and would like to see a replacement, please join the discussion thread at https://community.openastronomy.org/t/deprecating-sunpy-database/495.
+
+Deprecation of IO readers
+=========================
+``sunpy.io.cdf``, ``sunpy.io.file_tools`` and ``sunpy.io.jp2`` sub-modules have been deprecated, and will be removed in version 5.1.
+
+This because they are designed for internal use only, and removing it from the public API gives the developers more flexibility to modify it without impacting users.
