+++ sunpy/io/__init__.py
@@ -1,1 +1,1 @@
-from sunpy.io.file_tools import *
+from sunpy.io._file_tools import *
