+++ sunpy/timeseries/tests/test_timeseries_factory.py
@@ -125,7 +125,7 @@ def test_read_cdf_empty_variable():
     filename = Fido.fetch(result[0, 0])
 
     # Temporarily reset sunpy.io.cdf registry of known unit conversions
-    import sunpy.io.cdf as sunpy_cdf
+    import sunpy.io._cdf as sunpy_cdf
     known_units = sunpy_cdf._known_units
     sunpy_cdf._known_units = {}
